Fix: don't draw help window beyond screen height
The current computation for the drawing of the help window uses the
value `s->top + s->height` as the bottom edge of the screen, even
though the y values of the help strings being drawn are already
relative to the top of the screen. This results in some help strings
being drawn beyond the true edge of the screen, and therefore cut off.

Simply using `s->height` corrects this issue, and the help window gets
drawn properly even on multi-screen setups with differently sized
screens.

Signed-off-by: Antoine Busque <antoinebusque@gmail.com>